lm_backend/requirements.txt
Django>=4.1
djangorestframework
django-environ
python-dotenv
gunicorn
inference-sdk
google-generativeai
firebase-admin
pillow
requests
whitenoise
ultralytics>=8.0.20
torch>=2.0.0
opencv-python